Add and test stream package
The stream package implements a Stream type, DB resource and bucket.
Streams are user-facing interfaces to an underlying River type, which
manages nanomsg-compatible inproc Scalable Protocols sockets using
github.com/go-mangos/mangos.

This can be extended to use networked or even websocket SP sockets at a
later time, as they become needed.

The current types of River available use either the BUS protocol or the
PUB/SUB protocols.  Both of these have scaling issues and should be
revised at a later time if scaling becomes a high priority.  The stream
package should permit this without much difficulty.

When a River is loaded, it connects to other Rivers based on its
protocol and the River addresses stored in a bucket in the river bucket,
which should be named by the corresponding Stream's ID for Bus Rivers,
or the topic (WIP) for SUB Rivers.