Create and test REST /streams
Streams are REST entities with a User auth.  When a websocket
connection is requested (/streams/:id/start), a new BUS River is
created, binding to the user's websocket.  Messages sent on this ws
will be forwarded to all other users connected to the same River.

When the websocket is closed or the River socket is closed, the River is
removed atomically from the database.