Add WS subprotocol-based Bearer token auth middleware
Note that browser websocket clients cannot set their own headers,
except via the browser WebSocket API, which only offers the option to
set "subprotocols".  Subprotocols are inserted in the handshake request
as header values for the header 'Sec-WebSocket-Protocol'.

This commit adds an auth middleware for websockets which expects a
normal session token (retrieved by a GET on /tokens) to be passed as a
subprotocol of the format 'Bearer+<base64url encoding of token>'.  Note
that base64url permits padded and unpadded values; this middleware
expects an unpadded or "raw" version of the encoding [1].

The Stanford Javascript Crypto Library [2] offers a fast, small,
native Javascript implementation of unpadded base64url encoding, and
other decoding codecs.  To encode your token for this websocket
middleware, decode from padded base64 and encode the new value as
base64url.

[1] RFC-4648 section 5 https://www.ietf.org/rfc/rfc4648.txt
[2] http://bitwiseshiftleft.github.io/sjcl/